requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cffi==1.15.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==5.2.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.19.1
six==1.16.0
typing-extensions==3.10.0.2
yarl==1.7.0
youtube-dl==2021.6.6
